Add and use phpcs per-directory configs (#22117)
First, we have to use a forked version of PHPCS. squizlabs/PHP_CodeSniffer#3378
has been sitting there without response for six months now, and we're about
to really need the per-dir configuration to do the composer package
textdomain aliasing. Time to fork it.

Next, add projects for our fancy phpcs-filter, and the "ignorefile" library
that was created for it. Note no mirror repo is configured for the filter;
there's probably little point in publishing it until it doesn't require a
forked phpcs.

Then we use the new filter:
* Move all the exclusions to `.phpcsignore` files.
* Add per-dir phpcs configs to a few packages to reduce the need for
  `phpcs:ignore` comments
* Move the WordPress.WP.I18n `text_domain` config to per-project
  `.phpcs.dir.xml` files.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1596337634