Load the i18n-map in plugins (#22143)
For Jetpack, this will keep the packages' i18n working once we update
the packages to use aliasable domains instead of "jetpack".

For Backup and Boost, this will make the packages' i18n _start_ working
once we do that update (and once a version with `jetpack_vendor/` is
pushed to wporg svn so the translations exist).

When we add use of jetpack-composer-plugin to other plugins (e.g.
Search), we'll need to remember to do something similar to each of them
too.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1604112080