[Plugin] Backup: Add MyPlan component (#21759)
* Update Backup Segments section content.

Modify text shown on the section.
Add link to see all backups on Jetpack Cloud.
Remove the message to upgrade to realtime.

* Add API call to check plans and correct classnames

This commit corrects classnames, adds an API call to
check the plans details of the current site.
At this point, it only works with a single sub.

* Create new MyPlan component

This commit adds the MyPlan component and the logic
to detect Backup related purchases and list them.

* Add jp legacy plans to list of backup capable plans

* Migrate MyPlan API request and logic.
Migrate all the logic inside the component.
Modify the architecture to make it reusable.

* Add i8n fn to error message and default props
-Add translation function to the error message.
-Add default value to purchaseType prop.

* Change route name and use Fragment component
- Change route name to avoid using same route as JP
- Change empty tag for Fragment component

* Correct missmatched text domain

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1678452200