tools: Let check-changelogger-use.php detect uncommitted change entries (#23732)
If change entry files exist but aren't committed, the script should say
so instead of saying that they need to be added.

Further, when run from the pre-push hook, it would be nice if it would
offer to commit the entries.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2078595176