Stop updating to alpha versions in trunk (#38974)
As we do more standalone plugin releases in trunk, the need for PRs to
continually update version numbers for alpha versions (e.g. 1.2.3 →
1.2.4-alpha) is becoming onerous. After reviewing the current state of
monorepo tooling, we think that we can get away with having `jetpack
build` do these updates when building for the mirror repos while leaving
trunk at the old version until release time.

Removing the version bumping in trunk means we also don't have to have
the PR is up-to-date CI check show PRs as out of date after every
release of the affected projects.

The main negative potential effect we've identified with this is that
testing with `jetpack rsync` may reflect the old version numbers in UI
and logging. The Jetpack Autoloader should still be ok since it has
everything as "dev-trunk".

Note this does not remove the need for changelogger entries for all
changed projects in all PRs. We still need those for tooling to know
that something in the project has changed (and for changelogs to be
populated).

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/10582523341

Upstream-Ref: Automattic/jetpack@041ef4bb79eafeac2aa3d679c124fd5bfbc4f7dc